+++ test/samples/fallback-no-default/main.js
@@ -1,0 +1,4 @@
+var foo = require('./foo.js');
+
+assert.equal( foo.one, 1 );
+assert.equal( foo.two, 2 );